🚀 [Feat]: AppInitializer -> AuthInitializer로 명칭 변경 및 LoginInitializer 로직 병합
