--- src/shared/components/system/AppInitializer.tsx
@@ -1,19 +1,0 @@
-"use client";
-
-import { useAuthStore } from "@/shared/store/authStore";
-import { useEffect } from "react";
-
-interface AppInitializerProps {
-  isLoggedIn: boolean;
-}
-const AppInitializer = ({ isLoggedIn }: AppInitializerProps) => {
-  const { setLoggedIn } = useAuthStore();
-
-  useEffect(() => {
-    setLoggedIn(isLoggedIn);
-  }, [isLoggedIn]);
-
-  return null;
-};
-
-export default AppInitializer;
